INC-1165: ⏪️ Temporarily revert case note usage changes

+++ src/main/java/uk/gov/justice/hmpps/prison/api/model/CaseNoteUsageByBookingIdDto.kt
@@ -2,7 +2,7 @@ package uk.gov.justice.hmpps.prison.api.model
 import java.time.LocalDateTime
 
 data class CaseNoteUsageByBookingIdDto(
-  val bookingId: Long?,
+  val bookingId: Int?,
   val caseNoteType: String?,
   val caseNoteSubType: String?,
   val numCaseNotes: Int?,
